+++ runtime/vanilla/src/lib.rs
@@ -12,7 +12,6 @@ use orml_currencies::BasicCurrencyAdapter;
 use orml_traits::{parameter_type_with_key, DataProvider};
 use pallet_grandpa::fg_primitives;
 use pallet_grandpa::{AuthorityId as GrandpaId, AuthorityList as GrandpaAuthorityList};
-use primitives::{Amount, Balance, CurrencyId, Price};
 use sp_api::impl_runtime_apis;
 use sp_consensus_aura::sr25519::AuthorityId as AuraId;
 use sp_consensus_aura::SlotDuration;
@@ -279,11 +278,6 @@ impl orml_currencies::Config for Runtime {
     type WeightInfo = ();
 }
 
-parameter_types! {
-    pub const GetStableCurrencyId: CurrencyId = CurrencyId::USDT;
-    pub StableCurrencyFixedPrice: Price = 1;
-}
-
 impl pallet_loans::Config for Runtime {
     type Event = Event;
     type Currency = Currencies;
@@ -405,8 +399,6 @@ impl DataProvider<CurrencyId, TimeStampedPrice> for AggregatedDataProvider {
 impl pallet_prices::Config for Runtime {
     type Event = Event;
     type Source = AggregatedDataProvider;
-    type GetStableCurrencyId = GetStableCurrencyId;
-    type StableCurrencyFixedPrice = StableCurrencyFixedPrice;
     type FeederOrigin = EnsureRoot<AccountId>;
 }